Refactor HTML pages
